Adding some logging capabilities and some general improvements
