+++ src/common/object/jsc/methods.hpp
@@ -38,6 +38,7 @@ namespace method{
 
             return values[index];
         }
+        
     };
 
 };
