+++ src/test/test.sh
@@ -1,5 +1,6 @@
 clang++ -g -std=c++17 -Wall -I catch.hpp -I test_bed.hpp -I ../ \
     -I ../../vendor/realm-core/src/ \
+    -I includes/ \
     -I ../..//cmake-build-debug/vendor/realm-core/src/ \
     -I ../object-store/external/json \
     -framework Foundation -framework JavaScriptCore -o testing.o main.cpp catch.cpp \
